+++ dash_core_components_base/__init__.py
@@ -40,6 +40,8 @@ _current_path = _os.path.dirname(_os.path.abspath(__file__))
 _this_module = _sys.modules[__name__]
 
 async_resources = [
+    'datepicker',
+    'dropdown',
     'graph'
 ]
